requirements.txt
torch==2.3.1
einops==0.8.0
PyYAML==6.0.1
torchvision==0.18.1
scikit-learn==1.4.2
